refactor(plan-runner): delete plancheckpointer class and migrate resume-suite tests to fixture (resolves #2436)
