+++ tests/lib/orchestration/epic-plan-state-store.test.js
@@ -10,7 +10,7 @@ import {
   setPhase,
   write,
 } from '../../../.agents/scripts/lib/orchestration/epic-plan-state-store.js';
-import { PlanCheckpointer } from '../../../.agents/scripts/lib/orchestration/plan-runner/plan-checkpointer.js';
+import { PlanCheckpointer } from '../../fixtures/epic-plan-state-store.js';
 import { structuredCommentMarker } from '../../../.agents/scripts/lib/orchestration/ticketing.js';
 
 function createFakeProvider() {
